vision_model_containerzation/requirements.txt
torch
torchvision
timm
fastapi
uvicorn
pillow
pydantic
python-multipart
